search: don't offer an asset a search it abandoned
Clearing inprogress_by alone reopens the handback by a different route:
the abandoning asset is sitting on top of that search, so find_closest
ranks it first and find_next_search gives it straight back.

Exclude the requesting asset's abandonments from find_closest and
oldest_queued_for_asset_type. oldest_queued_for_asset is deliberately
left alone - queueing a search for one named asset is an explicit
operator assignment that should override an earlier abandonment.

Other assets are still offered the search; that is the reassignment.

The abandon tests move to search/test_abandon.py: search/tests.py was
over pylint's 1000 line limit, and per-feature test modules are already
the pattern here (search/test_views_additional.py, mission/test_asset_command.py).
