+++ search/tests.py
@@ -819,127 +819,3 @@ class SearchClosedMissionTestCase(TestCase):
         response = self.smm.client1.post(f'/search/{self.search.search_id}/finished/', data={'asset_id': self.asset.pk})
         self.assertEqual(response.status_code, 403)
         self.assertIsNone(self.search.as_object().completed_at)
-
-
-class SearchAbandonModelTestCase(TestCase):
-    """
-    Tests for Search.abandon()
-    """
-    def setUp(self):
-        self.smm = SMMTestUsers()
-        assets = AssetsHelpers(self.smm)
-        self.searches = SearchHelpers(self.smm)
-        self.asset_type = assets.create_asset_type()
-        self.asset = assets.create_asset(name='asset_a', asset_type=self.asset_type)
-        self.other_asset = assets.create_asset(name='asset_b', asset_type=self.asset_type)
-        self.mission = MissionFunctions(self.smm).create_mission('abandon mission')
-        self.mission.add_asset(self.asset)
-        self.mission.add_asset(self.other_asset)
-        self.poi = GeoTimeLabel.objects.create(geo=Point(172.5, -43.5), created_by=self.smm.user1, label='datum', geo_type='poi', mission=self.mission.get_object())
-
-    def create_inprogress_search(self, asset=None):
-        """
-        Create a search and put it in progress by an asset
-        """
-        if asset is None:
-            asset = self.asset
-        search = self.searches.create_sector(self.poi, 200, self.asset_type).as_object()
-        self.assertTrue(search.set_inprogress_by(asset, self.smm.user1))
-        return search
-
-    def test_abandon_clears_inprogress_and_queue(self):
-        """
-        Abandoning releases the asset and takes the search out of the queue
-        """
-        search = self.create_inprogress_search()
-        Search.objects.filter(pk=search.pk).update(queued_at=timezone.now(), queued_for_asset=self.asset)
-        search.refresh_from_db()
-
-        self.assertTrue(search.abandon(self.asset, self.smm.user1))
-
-        search.refresh_from_db()
-        self.assertIsNone(search.inprogress_at)
-        self.assertIsNone(search.inprogress_by)
-        self.assertIsNone(search.queued_at)
-        self.assertIsNone(search.queued_for_asset)
-
-    def test_abandon_returns_search_to_all_waiting(self):
-        """
-        An abandoned search is available to be picked up again
-        """
-        search = self.create_inprogress_search()
-        mission = self.mission.get_object()
-        self.assertNotIn(search, Search.all_waiting(mission))
-
-        search.abandon(self.asset, self.smm.user1)
-
-        self.assertIn(search, Search.all_waiting(mission))
-
-    def test_abandon_records_the_asset(self):
-        """
-        The abandoning asset is recorded so it can be excluded later
-        """
-        search = self.create_inprogress_search()
-
-        search.abandon(self.asset, self.smm.user1)
-
-        self.assertEqual(list(search.abandoned_by.all()), [self.asset])
-
-    def test_abandon_records_timeline_entry(self):
-        """
-        Abandoning a search is visible in the mission timeline
-        """
-        search = self.create_inprogress_search()
-
-        search.abandon(self.asset, self.smm.user1)
-
-        entries = TimeLineEntry.objects.filter(mission=self.mission.get_object(), event_type='sab')
-        self.assertEqual(entries.count(), 1)
-        self.assertIn(str(self.asset), entries[0].message)
-
-    def test_abandon_rejects_other_asset(self):
-        """
-        An asset cannot abandon a search another asset is conducting
-        """
-        search = self.create_inprogress_search()
-
-        self.assertFalse(search.abandon(self.other_asset, self.smm.user1))
-
-        search.refresh_from_db()
-        self.assertEqual(search.inprogress_by, self.asset)
-        self.assertEqual(TimeLineEntry.objects.filter(event_type='sab').count(), 0)
-
-    def test_abandon_rejects_not_inprogress_search(self):
-        """
-        A search nobody is conducting cannot be abandoned
-        """
-        search = self.searches.create_sector(self.poi, 200, self.asset_type).as_object()
-
-        self.assertFalse(search.abandon(self.asset, self.smm.user1))
-
-    def test_abandon_rejects_completed_search(self):
-        """
-        A completed search cannot be abandoned
-        """
-        search = self.create_inprogress_search()
-        search.completed_at = timezone.now()
-        search.completed_by = self.asset
-        search.save()
-
-        self.assertFalse(search.abandon(self.asset, self.smm.user1))
-
-        search.refresh_from_db()
-        self.assertEqual(search.inprogress_by, self.asset)
-
-    def test_abandon_rejects_deleted_search(self):
-        """
-        A deleted search cannot be abandoned
-        """
-        search = self.create_inprogress_search()
-        Search.objects.filter(pk=search.pk).update(deleted_at=timezone.now(), deleted_by=self.smm.user1)
-        search.refresh_from_db()
-
-        self.assertFalse(search.abandon(self.asset, self.smm.user1))
-
-        search.refresh_from_db()
-        self.assertEqual(search.inprogress_by, self.asset)
